requirements.txt
streamlit
opencv-python
ultralytics
pandas
pillow